Add Server Create key_name
Following the implemented endpoints for the keypairs, we 've added
functionality for the keypair injection onto the VM creation. Once a
user wants to create a new VM, he can provide a `key_name` which will be
used for obtaining a stored keypair. The contents of this keypair will
be injected to the new machine, using an extension in the ganeti client